move pages into their own directory
