+++ src/pages/HappHosting/HappHosting.connector.js
